fix: Mypage에서 평점의 소숫점이 안나오는 것 수정, 자기 자신 투표 막음

+++ server/src/main/java/com/team_60/Mocco/member/mapper/MemberMapper.java
@@ -37,8 +37,8 @@ public interface MemberMapper {
     MyInfo memberPatchDtoToMyInfo(MemberDto.Patch dto);
 
     default MemberDto.Response memberToMemberResponseDto(Member member){
-        int evaluation = member.getMyInfo().getEvaluationNumber() == 0 ?
-                0 : member.getMyInfo().getEvaluationTotal() / member.getMyInfo().getEvaluationNumber();
+        float evaluation = member.getMyInfo().getEvaluationNumber() == 0 ?
+                0 : (float) member.getMyInfo().getEvaluationTotal() / (float) member.getMyInfo().getEvaluationNumber();
 
         List<String> github_repository_list = new ArrayList<>();
         github_repository_list.add(member.getMyInfo().getGithubRepository1());
